Re-organize and improve the smart fetch code
